refactor: move string tests
